% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additions_vc_theory.R
\name{check_objset_sufg_candidate}
\alias{check_objset_sufg_candidate}
\title{Check if subset is enlargable to a sufg premise of size s}
\usage{
check_objset_sufg_candidate(subset, context, s)
}
\arguments{
\item{subset}{is the envisaged subset given as a 0-1-vector.}

\item{context}{is the underlying context.}

\item{s}{is the envisaged cardinality of the set to which the set 'subset'

can or can not be extended.}
}
\value{
A list with entry result which is TRUE if the set can be extended,
otherwise it is FALSE. The entry set is the set to which 'subset' can be
enlarged, given as a 0-1-vector.
}
\description{
'check_objset_sufg_candidate' checks if a given set of objects
is extendable to a ufg-premise of size s.
If one wants to check if a given set'subset' is itself an ufg-premise, then
one can simply set s=|subset|.
Note that we assume here that the underlying context 'context' represents the
whole data space (i.e., the small context is the large context, thus the term
sufg (s for small) instead of ufg)
}
\examples{
context <- matrix(c(
  1, 1, 0, 0, 1, 0, 1,
  1, 0, 0, 0, 0, 1, 0,
  0, 1, 0, 0, 0, 0, 0,
  0, 1, 0, 1, 0, 0, 0,
  0, 1, 0, 1, 0, 0, 0,
  0, 0, 1, 1, 1, 0, 0,
  0, 0, 0, 0, 0, 1, 0,
  0, 0, 0, 0, 0, 1, 0,
  0, 1, 0, 1, 0, 0, 0,
  1, 0, 1, 0, 0, 1, 0
), nrow = 10, ncol = 7, byrow = TRUE)

fc <- fcaR::FormalContext$new(t(context))

# Implikationen:

fc$find_implications()
fc$implications

# aus der kanonischen Implikationsbasis ist die ufg Basis schwer abzuleiten,
# einfacher ist es, denn Begriffsverband anzuschauen. Z.b. ist es einfach
# abzulesen, dass {A1,A6,A10} -> {A1,A2,.., A10}
# gilt, da das Infimum von A1,A6,A10 das kleinste Elemente ist und damit
# unter Gegenstaenden liegt (ACHTUNG: Im Bild wird der Verband von t(context)
# und nicht context berechnet (dann werden im plot die Begriffe mit den
# entrechenden Gegenstaenden A1,...A10 bezeichnet und man kann die
# Gegenstandsimplikationen einfach ablesen.





fc$find_concepts()
fc$concepts$plot()
subset <- rep(0, 10)
subset[c(7)] <- 1
check_objset_sufg_candidate(subset, context, 1)

# TRUE
# passt denn {A7} -> {A8,A2,A10,A7} ist ufg Implikation
check_objset_sufg_candidate(subset, context, 2)
# TRUE
# passt denn z.B. {A7,A3} -> {A1,... A10} ist ufg Implikation
# objset_is_sufg_candidate(subset,context,3)
# FALSE
# scheint ebenfalls zu passen
subset <- rep(0, 10)
subset[c(7, 8)] <- 1
check_objset_sufg_candidate(subset, context, 2)
# FALSE passt

subset <- rep(0, 10)

check_objset_sufg_candidate(subset, context, 3)

# $result
# [1] TRUE
# $set
# [1] 1 0 0 0 0 1 0 0 0 1

# passt, denn die Leere Menge ist zu der ufg Praemisse {A1,A6,A10}
# erweiterbar (siehe oben)

}
