% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additions_vc_theory.R
\name{compute_extent_vc_dimension}
\alias{compute_extent_vc_dimension}
\title{Computes a MILP model for the computation of the VC-dimension of a formal
context.}
\usage{
compute_extent_vc_dimension(context, additional_constraint = TRUE)
}
\arguments{
\item{context}{is the underlying formal context.}

\item{additional_constraint}{is an additional constraint within the MILP
formulation for computing th VC dimension. If set to 'TRUE' then it is
explicitly demanded that the number of the objects and the number of the
attributes that are involved in the maximum contranominal scale (which is
directly connected to a shatterable subset of maximal cardinality, cf., TODO
) are identical.}
}
\value{
a MILP model that can be optimized with 'gurobi(.)'. The VC dimension
is then given bei the maximum value ('$objval') of the optimized model.
The contranominal scale that represents the shatterable set of maximal
cardinality can be obtained from the model with '$x'. This vector is a 0-1
vector of length number_objects + number_attributes which represents
the objects concatenated with the attributes.
}
\description{
'compute_extent_vc_dimension' builds a MILP model for the
computation of the Vapnik-Chervonenkis dimension of the family of all extents
of the concept lattice of a formal context. The variables of the MILP model
are a 0-1 vector of length number_objects + number_attributes. The first
'number_objects' variables model the objects (variable x_i is 1 if object
i is in the contranominal scale that represents the shatterable set of
maximal cardinality) and the last 'number_attributes ' variables represent
the attributes (variable x_{j+number_objects} is 1 if attribute j is in the
contranominal scale that represents the shatterable set of maximal
cardinality).
}
\examples{


chevron <- rbind(
  c(1, 1, 0, 0, 0, 0), c(0, 1, 0, 0, 0, 0), c(1, 1, 1, 0, 1, 1),
  c(0, 1, 0, 1, 0, 1), c(0, 0, 0, 0, 1, 1), c(0, 0, 0, 0, 0, 1)
)
oofos:::plot_relation(chevron)

model <- compute_extent_vc_dimension(chevron)
vc <- gurobi:::gurobi(model, list(outputflag = 0))
vc$objval

# [1] 2  Vc dimenion of the family of all intersections of pronicpal filter
# should be 2

which(vc$x[(1:6)] == 1)
# [1] 2 5  For example the set of elements 2 and 5 can be shattered by the
# extents of the context (There are more shatterable sets of size 2)


model <- compute_extent_vc_dimension(1 - chevron)

# Here, we analyze the contraordinalscale for the chevron (1-chevron!) .
# In this case the extents are exactly the upsets
# cf., Mathematische Strukturen Ws19/20 Uebung 3, Aufgabe 2 (Oberhalbmengen
# sind genau die Umfaenge der Kontraordinalskala)
# In this case the VC dimension is exactly the width of the contraordinal
# context, cf., Schollmeyer 2017 TR 209, Definition & Proposition 2, p. 31).
vc <- gurobi::gurobi(model, list(outputflag = 0))

vc$objval

# [1] 3 VC dimension should be 3


which(vc$x[(1:6)] == 1)

# [1] 1 4 5 For example the set of elements 1 4 and 5 is shatterabel /
# an antichain

}
