% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{compute_objective}
\alias{compute_objective}
\title{Compute the objective vector for linear optimization}
\usage{
compute_objective(
  dat,
  target,
  target_class,
  weights = rep(1, length(dat[[target]]))
)
}
\arguments{
\item{dat}{is the underlying data set}

\item{target}{is a character that specifies which variable of the dataset dat
serves as target.}

\item{target_class}{specifies for which value of the target variable the
relative frequencies are considered.}

\item{weights}{is a possible weighing vector for the target variable}
}
\value{
the vector of the objective function, namely
w^x-w^y where w^x are the relative frequencies of the target taking the value
'target_class' (the denominator is the number of all data points with value
'target_value'). More concretely w^x is 1/n^x if the coressponding data point
takes the value 'target_value' where n^x s the number of data points with
value 'target_value'. Otherwise w^x is zero. Analoguously, w^y = 1/n^y if the
corresponding data point has not the value 'target_value' where n^y is the
number of data points that have not the value 'target_value'. And w^y = 0
otherwise.
}
\description{
'compute_objective computes the objective vector that
corresponds to the spremum type objective function

max < w^x,m > - < w^y,m >

where m ranges over all indicator functions of the considered
family of sets, cf, Schollmeyer et al. 2017, p.8f, p.20
}
\references{
Schollmeyer, G., Jansen, C., Augustin, T. (2017):
Detecting stochastic dominance for poset-valued random variables as an
example of linear programming on closure systems. Technical Report 209,
Department of Statistics, LMU Munich.
}
