% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starshaped_sets.R
\name{compute_quotient_order}
\alias{compute_quotient_order}
\title{Compute the quotient order of a quasiorder}
\usage{
compute_quotient_order(incidence)
}
\arguments{
\item{incidence}{is the incidence matrix of the quasiorder}
}
\value{
the incidence matrix of the quotient order
}
\description{
'compute_quotient_order' computes the quotient order of a
quasiorder by deleting equivalent 'duplicates' from the given incidence
matrix of the quasiorder. if 'incidence' is not a quasiorder, then first the
transitive reflexive hull is computet.
}
