% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starshaped_sets.R
\name{compute_starshaped_distr_test}
\alias{compute_starshaped_distr_test}
\title{Compute a statistical significance test for a starshaped subgroup discovery}
\usage{
compute_starshaped_distr_test(ssd_result, n_rep = 1000, plot_progress = TRUE)
}
\arguments{
\item{ssd_result}{is the result of a starshaped subgroup discovery that was
computed with the function 'discover_starshaped_subgroups'.}

\item{n_rep}{is the number of permutations in the observation-randomization
test.}

\item{plot_progress}{if TRUE (default) then the progress of the
observation-randomization test is plot by drawing the empirical distribution
function of the currently computed resampled objective values. Additionally,
nonparametric and parametric estimates of the p-value based on the currently
computed resamples are printed. TODO : beta genauer}
}
\value{
A list with entries 'objvalues': the resampled objective values,
'p_value'  the (nonparametrically) estimated p-value of the
observation-randomization test, 'p_value_parametric': a parametric estimate
of the p-value based on a beta-approximation of the distribution of the
resampled objective values.
}
\description{
'compute_starshaped_distr_test' performs a statistical
significance test to assess if the largest differences of relative
frequencies over all starshaped subgroups that was found in a starshaped
subgroup discovery is statistically significantly diefferent from zero.
}
