% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starshaped_sets.R
\name{compute_stylized_betweenness}
\alias{compute_stylized_betweenness}
\title{Compute the canoncical (stylized) betweennes relation between objects of a formal context}
\usage{
compute_stylized_betweenness(
  g,
  h,
  i,
  context,
  attribute_weights = colMeans(context)
)
}
\arguments{
\item{g}{is the object g represented by a 0-1 vector of attributes.}

\item{h}{is the object h represented by a 0-1 vector of attributes.}

\item{i}{is the object i represented by a 0-1 vector of attributes.}

\item{context}{is the underlying context.}

\item{attribute_weights}{is the weight vector for the attributes.}
}
\value{
a ternary fuzzy relation given by an array of dimension
n x n x n where n is the number of rows of the context. Higher values
of an entry correspond to a larger degree of betweenness.
}
\description{
'compute_stylized_betweenness' computes the canonical (stylized)
betweennes relation between objects of a formal context: Given three
objects g,h,i of a formal context we say that h lies betweeen g and i iff


Psi({g}) cap Psi({i}) subseteq Psi({h}).


Additionally we say in a stylized
manner that h lies betweeen g and i iff
Psi({g}) cap Psi({i}) subseteq Psi({h}) is almost true with the exception of
some attributes of Psi({g}) cap Psi({i}) that do not belong to Psi({h}) but
that are not so 'important' in the sense that there are not too much objects
that have all attributes in (Psi({g}) cap Psi({i})) -  Psi({h}) but that have
not all attributes in Psi({h}). Concretely the stylized betweennes is
quantified by the maximum of the weights 'attribute_weights' associated
to these attributes. (The weights could be defined for example as the cloumn
means of the underlying formal context).
}
