% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additions_vc_theory.R
\name{compute_sufg_dimension}
\alias{compute_sufg_dimension}
\title{Compute the (s)ufg-dimension of a formal context}
\usage{
compute_sufg_dimension(context, additional_constraint = TRUE)
}
\arguments{
\item{context}{is the underlying context.}

\item{additional_constraint}{If set to TRUE, then an additional constraint
for the MILP formulation is added which sometimes speeds up computation.}
}
\value{
A MILP model that can be optimized with 'gurobi(.)'. The optimal
value ('$objval') of the optimized model then corresponds to the (s)ufg
dimension of the context.


context' (s bezieht sich auf
}
\description{
'compute_sufg_dimension' computes the (s)ufg-dimension of a
formal context, i.e., the maximal cardinality of a (s)ufg-premise. Here we
assume that the underlying ground space is represented by the formal context
'context', i.e. the formal context representing the whole space we are
thinking in (the 'large context') is identical to the given context 'context'
that at the same time rpresents the context representing the observed data
points (the 'small context'). Thus the addition s (for small).
}
