% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additions_vc_theory.R
\name{compute_ufg_depth}
\alias{compute_ufg_depth}
\title{Compute ufg-depth for dataset of partial orders}
\usage{
compute_ufg_depth(
  data_context,
  evaluation_context,
  ufg_list,
  counts,
  n = sum(counts)
)
}
\arguments{
\item{data_context}{is the context of observed partial orders given as a
complemented context. Usually the context data_context should be supplied as
a context without duplicates. (the the integer vector counts of size
nrow(data_context), see below, then specifies, how often every partial order
was observed.}

\item{evaluation_context}{is the set of partial orders for which the ufg
depth values should be computed, given as a complemented context.
ufg_list is a list of indices (w.r.t 'data_context') indicating which
sets are ufg-sets. This list can be computed by the function
'enumrate_ufg-premises'}

\item{counts}{Usually the context data_context should be supplied as a
context without duplicates. The integer vector counts of size
nrow(data_context) then specifies, how often every partial order was
observed.}
}
\value{
list with depth values and the total number of ufg-premises
}
\description{
'compute_ufg_depth' computes the ufg-depth of a data set of
partial orders given as a complemented context
}
