% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additions_order_theory.R
\name{compute_width}
\alias{compute_width}
\title{Compute the width of a partiallyordered set}
\usage{
compute_width(incidence)
}
\arguments{
\item{incidence}{square incidence matrix with 0-1 entries}
}
\value{
a list with entry width that gives the width of the poset.
}
\description{
'compute_width' computes the width of partially ordered set,
i.e., the maximal cardinality of an antichain. (An antichain is a set of
pairwise incomparable elements.)
For computing the width is done with a maximum matching formulation, see
\url{https://mathoverflow.net/questions/189161/fastest-algorithm-to-compute-the-width-of-a-poset}
}
