% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starshaped_sets.R
\name{discover_starshaped_subgroups}
\alias{discover_starshaped_subgroups}
\title{Perform a starshaped subgroup discovery}
\usage{
discover_starshaped_subgroups(
  stylized_betweenness,
  objective,
  local_vc_dimension,
  params = list(Outputflag = 0)
)
}
\arguments{
\item{stylized_betweenness}{is the fuzzy betweennness relation.}

\item{objective}{is the (linear) objective function to optimize given as an
objective vector of the same length as the number of objects.}

\item{local_vc_dimension}{Is the local VC dimension: Given a centerpoint
c, the VC dimension of the subfamily of all starshaped sets with centerpoint
c is controlled with this parameter. If local_vc_dimension is set to Inf,
then the complexity of the family of all starshaped sets is not reduced at
all.}

\item{params}{is a list with further arguments that are passed to the gurobi
optimizer. By default the outputflag of the gurobi solver is set to 0 (this
means that not optimization details are printed during the optimization).}
}
\value{
a list with the following entries: TODO
}
\description{
'discover_starshaped_subgroups' performs a starshaped subgroup
discovery, see Schollmeyer et al. 2023. An objective function (e.g., the
Shapiro- Piatetsky quality function) is maximized over the family of all
subgroups that are starshaped w.r.t. a 'fuzzy' betweennes relation

TODO : explain VC trimming...
}
