% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additions_vc_theory.R
\name{enumerate_ufg_premises}
\alias{enumerate_ufg_premises}
\title{Enumerate all ufg premises of a complemented poset context}
\usage{
enumerate_ufg_premises(whole_context, n_row_context, print_progress = TRUE)
}
\arguments{
\item{whole_context}{is the whole context that describes the whole space of
complemented partial orders.

For the implementation with 'test_ufg_porder' as a test function for checking
ufg-candidates it is not needed to supply the whole context of all partial
orders, instead it is enough to supply only the orders obtained in the data
sample.

For the implementation with 'test_explicitly_ufg_p_order' as a test function
for checking ufg-candidates it iis needed to supply the context of all
partial orders.

The current implementation is ...}

\item{n_row_context}{is the number of objects in the observed data sample.
It is assumed that the first 'n_row_context' rows of the context
'whole_context' represent the observed partial orders.}

\item{print_progress}{If TRUE, the progress of the enumeration will printed.}
}
\value{
A list of all ufg premises given as a vector of indices (w.r.t. the
first 'n_row_context' rows of the context 'whole_context').
}
\description{
'enumerate_ufg_premises' enumerates all ufg premises for a set
of partial orders given as a complemented context
}
