% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starshaped_sets.R
\name{fit_ks_distribution}
\alias{fit_ks_distribution}
\title{Approximate the distribution of a Kolmogorov-Smirnov types test statistic by
the (noncentral) beta distribution}
\usage{
fit_ks_distribution(objvalues, plot_result = FALSE)
}
\arguments{
\item{objvalues}{sampled values of a Kolomogorov Sminrov type test statistic}

\item{plot_result}{If true the empirical distribution function of the
'objvalues'-values together with the cdf of the fitted beta distribution is
plotted.
description}
}
\description{
TODO
}
\references{
Zhang, J., Wu, Y. Beta Approximation to the Distribution of
Kolmogorov-Smirnov Statistic. Annals of the Institute of Statistical
Mathematics 54, 577–584 (2002). https://doi.org/10.1023/A:1022463111224
}
