% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additions_vc_theory.R
\name{get_weighted_representation}
\alias{get_weighted_representation}
\title{compute weighted representation of a data matrix}
\usage{
get_weighted_representation(x, y = rep(1, dim(x)[1]))
}
\arguments{
\item{x}{TODO}

\item{y}{TODO}
}
\description{
computes weighthed representation of a data matrix x with
duplicated rows, returns unique(x) together with counts: how often appears
the column, mean_y: mean of y in the set of the duplicated columns
WARNING: the order of the rows generally differs from the order one would get
by applying the function 'unique()'.
}
