% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{optimize_on_context_extents}
\alias{optimize_on_context_extents}
\title{Optimize a linear function on the family of all extents of a formal context}
\usage{
optimize_on_context_extents(
  context,
  gen_index = (1:nrow(context)),
  objective,
  binary_variables = "afap"
)
}
\arguments{
\item{context}{is the underlying formal context.}

\item{gen_index}{is an index vector that indicates the set of objects which
will be considered as the obects that generate the formal concept lattice
over which concept extents the optimization is done. By default it is the
vector of the indices of all objects of the context. One can show that the
maximization over the concept lattice that is generated only by all objects
with a positive objective value 'objective' will lead to the same
optimization problem as if one would optimize over the concept lattice
generated by all objects reference: TODO}
}
\description{
'optimize_on_context_extents' maximizes a linear function over
 all indicator sets of all extents of a given formal context.
}
