% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additions_fca.R
\name{plot_relation}
\alias{plot_relation}
\title{Plot a binary relation}
\usage{
plot_relation(incidence)
}
\arguments{
\item{incidence}{The incidence relation of the partial order.}
}
\value{
The drawing of the partial order in the form of the Hasse diagramm of
the Dedekind-MacNeille completion of the given partial order.
}
\description{
'plot_relation' plots a binary relation(e.g., a partial order)
by drawing the Hasse diagramm of its Dedekind–MacNeille completion. Note
that the plotted relation needs not to be a partial order, it can be an
arbitrary relation.
}
\examples{

# plot an interval order
upper <- (1:10)
lower <- upper - runif(10)
interval_order <- array(0, c(10, 10))
for (k in (1:10)) {
  for (l in (1:10)) {
    interval_order[k, l] <- (upper[k] <= lower[l])
  }
}
oofos:::plot_relation(interval_order)

# plot a quasiorder
q_order <- array(1 * (runif(100) >= 0.9), c(10, 10))
diag(q_order) <- 1
q_order[1, 2] <- q_order[2, 1] <- 1
q_order <- oofos:::compute_transitive_hull(q_order)
oofos:::plot_relation(q_order)

# plot the Chevron
chevron <- rbind(
  c(1, 1, 0, 0, 0, 0), c(0, 1, 0, 0, 0, 0), c(1, 1, 1, 0, 1, 1),
  c(0, 1, 0, 1, 0, 1), c(0, 0, 0, 0, 1, 1), c(0, 0, 0, 0, 0, 1)
)
oofos:::plot_relation(chevron)

}
