% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additions_vc_theory.R
\name{sample_ufg_K_objset_recursive}
\alias{sample_ufg_K_objset_recursive}
\title{##TODO}
\usage{
sample_ufg_K_objset_recursive(
  context,
  K,
  N = rep(nrow(context), K),
  threads = 1,
  model = NULL
)
}
\description{
##TODO
}
