% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additions_vc_theory.R
\name{enumerate_ufg_premises}
\alias{enumerate_ufg_premises}
\title{Enumerate all ufg premises of a complemented poset context}
\usage{
enumerate_ufg_premises(
  whole_context,
  n_row_context,
  n_ufgs = 1e+07,
  print_progress = TRUE
)
}
\arguments{
\item{whole_context}{is the whole context that describes the whole space of complemented
partial orders.}

\item{n_row_context}{is the number of objects in the observed data sample.
It is assumed that the first 'n_row_context' row of the ccontext
'whole_context' represent the observed partial orders.}

\item{n_ufgs}{is an upper bound for the number of ufg premises. This is needed to
allocate enough memory for the result.}

\item{print_progress}{If TRUE, the progress of the enumeration will printet.}
}
\value{
Amtraix where every row represents an ufg premise by giving the
indices of the partial orders from the context 'whole_context'. Since
different ufg premises can have different cardinalities, for ufg premises
with smaller cardinalities the corresponding vectors are filled with zeros
}
\description{
TODO
}
