% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{optimize_on_context_extents}
\alias{optimize_on_context_extents}
\title{Optimize a linear function on the family of all extents of a formal context}
\usage{
optimize_on_context_extents(
  context,
  gen_index = seq_len(nrow(context)),
  objective,
  binary_variables = "afap"
)
}
\arguments{
\item{context}{is the underlying formal context.}

\item{gen_index}{is an index vector that indicates the set of objects which
will be considered as the obects that generate the formal concept lattice
over which concept extents the optimization is done. By default it is the
vector of the indices of all objects of the context. One can show that the
maximization over the concept lattice that is generated only by all objects
with a positive objective value 'objective' will lead to the same
optimization problem as if one would optimize over the concept lattice
generated by all objects reference: TODO}

\item{objective}{is a vector of length nrow(context) that represents the linear
function that is optimized.}

\item{binary_variables}{one of the characters "afap", "sd", "allgen", "al"
(short for 'as few as possible', 'smallest dimension', 'all generating
objects' or 'all variables'. This determines which variables are set to
binary ones. Depending on this, the speed of the optimization can vary very
strongly.

TODO genauer beschreiben}
}
\value{
a model that can be optimized by calling gurobi(...).
}
\description{
'optimize_on_context_extents' maximizes a linear function over
 all indicator sets of all extents of a given formal context.
}
\examples{
\dontrun{
# needs Package CDM (Cognitive Diagnosis Modelling)
dat <- na.omit(CDM::data.timss07.G4.lee$data)
context <- dat[,-(1:3)]
target <- compute_objective (dat,"girl",0)
model <- optimize_on_context_extents(context,(1:344),target)
result <- gurobi::gurobi(model)
result$objval}

}
